requirements.txt
pandas
requests
openpyxl
open-webui

gdown==5.2.0
docker==7.1.0
pydub==0.25.1
ollama==0.4.7
moviepy==1.0.3
selenium==4.28.1
screeninfo==0.8.1
open-webui==0.5.16
platformdirs==4.3.6
pytest-docker==3.1.2
selenium-wire==5.1.0
cryptography==44.0.0
openai-whisper==20240930
importlib_metadata==8.4.0
importlib_resources==6.5.2

wheel
gradio
gspread
colorama
selenium
requests
setuptools
pycryptodome

pyyaml>=6.0
aiohttp>=3.8.1
blinker==1.7.0
cython>=0.29.30
packaging>=23.1
anyascii>=0.3.0
fsspec>=2023.6.0

google-auth-httplib2
google-auth-oauthlib
undetected-chromedriver
google-api-python-client

flask
flask_cors
torch